+++ CHANGELOG.md
@@ -54,6 +54,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * New `tracked_remote_branches()` and `untracked_remote_branches()` revset
   functions can be used to select tracked/untracked remote branches.
 
+* The `file()` revset function now accepts fileset as argument.
+
 ### Fixed bugs
 
 * `jj diff --git` no longer shows the contents of binary files.